CXP-1125: Move ApiAuthenticationEvent to correct namespace
